requirements.txt
pyqt5
pyusb
psutil
nvidia-ml-py3
